requirements.txt
appdirs==1.4.3
certifi==2016.2.28
contextlib2==0.5.4
dill==0.2.1
eventlet==0.19.0
falcon==1.0.0
Flask==0.10.1
funcsigs==1.0.2
future==0.15.2
greenlet==0.4.10
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.2
MarkupSafe==0.23
mock==2.0.0
numpy==1.12.1
oauth==1.0.1
oauthlib==0.7.2
packaging==16.8
pbr==2.0.0
peewee==2.8.1
py==1.4.31
pymongo==3.3.0
pyparsing==2.2.0
pytest==3.0.3
pytest-mock==1.6.0
python-mimeparse==1.5.2
python-telegram-bot==4.3.2
raven==5.27.1
redis==2.10.5
requests==2.10.0
simplejson==3.8.0
six==1.10.0
urllib3==1.16
Werkzeug==0.11.10
